\name{getAllParts}
\alias{getAllParts}
\title{
Return vector or matrix of linked entities from vector of long strings with
entities separated by a common separator.
}
\description{
Take a collection of linked pairs in a common string separated by a common
separator and return a matrix of individual entites.
}
\usage{
getAllParts( fullfnames , sep = "-" , retLen = FALSE )
}
\arguments{
  \item{fullfnames}{
    Vector of long strings to be separated.
  }
  \item{sep}{
    Separator.
  }
  \item{ret_len}{
    Return lengths of entity pairs.
  }
}
\details{
%%
}
\value{
%% 
}
\author{
Bin Zhang
}
\examples{
## To be added
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ split }
